Added app_factory conneting routes
