+++ services/api-gateway/src/simcore_service_api_gateway/endpoints_user.py
@@ -1,18 +1,24 @@
+from fastapi import APIRouter, Depends, Security
+
+from .auth import get_current_active_user, get_current_user
+from .schemas import User
+
+router = APIRouter()
 
 
-
-@app.get("/users/me/", response_model=User)
+@router.get("/users/me/", response_model=User)
 async def read_users_me(current_user: User = Depends(get_current_active_user)):
     return current_user
 
 
-@app.get("/users/me/projects/")
+@router.get("/users/me/projects/")
 async def list_own_projects(
     current_user: User = Security(get_current_active_user, scopes=["projects"])
 ):
     return [{"project_id": "Foo", "owner": current_user.username}]
 
 
-@app.get("/status/")
+@router.get("/status/")
 async def read_system_status(current_user: User = Depends(get_current_user)):
+    print(current_user)
     return {"status": "ok"}
