+++ services/api-gateway/src/simcore_service_api_gateway/main.py
@@ -5,37 +5,25 @@ from pathlib import Path
 from fastapi import FastAPI
 from tenacity import Retrying
 
-from .__version__ import api_version, api_version_prefix
+from . import app_factory
 from .config import is_testing_enabled
 from .db import create_tables, pg_retry_policy, setup_engine, teardown_engine
-from .endpoints import diagnostics, samples
 from .utils.remote_debug import setup_remote_debugging
 
 current_dir = Path(sys.argv[0] if __name__ == "__main__" else __file__).resolve().parent
 
 log = logging.getLogger(__name__)
 
+app: FastAPI = app_factory.create()
 
-# APPLICATION
-app = FastAPI(
-    debug = is_testing_enabled,
-    title="Public API Gateway",
-    description="Platform's API Gateway for external clients",
-    version=api_version,
-    openapi_url=f"/api/{api_version_prefix}/openapi.json"
-)
-
-# ROUTES ----
-app.include_router(diagnostics.router, tags=['diagnostics'])
-app.include_router(samples.router, tags=['samples'], prefix=f"/{api_version_prefix}")
-
-# NOTE: yaml.safe_dumps(app.openapi())
 
 # EVENTS ----
+# TODO: move to app_factory.py
 @app.on_event("startup")
 def startup_event():
-    log.info( "Application started")
+    log.info("Application started")
     setup_remote_debugging()
+
 
 @app.on_event("startup")
 async def start_db():
@@ -66,3 +54,4 @@ async def shutdown_db():
 
 # with open(current_dir / "openapi.json", "wt") as fh:
 #     json.dump(app.openapi(), fh, indent=2)
+# NOTE: yaml.safe_dumps(app.openapi())
